feat: add ReduceStrict method
